requirement.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.8
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
openai==2.30.0
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlite-vec==0.1.9
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
xxhash==3.6.0
zstandard==0.25.0
